# Use a custom swiss-like seriesHashmap impl.