--- a/tsdb/series_hashmap.go
+++ b/tsdb/series_hashmap.go
@@ -10,6 +10,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+//
 // Provenance-includes-location: https://github.com/dolthub/swiss/blob/f4b2babd2bc1cf0a2d66bab4e579ca35b6202338/map.go
 // Provenance-includes-license: Apache-2.0
 // Provenance-includes-copyright: Copyright 2023 Dolthub, Inc.