Right now `seriesHashmap` is a struct with two maps, one for the happy case, and one for the hash collisions we'd have.

Let's ignore for a moment the hash collisions: we have a hash, and we put it in a map so we have to hash the hash. Same for the retrievals.

What if we just used our hash as the internal hash for the hashmap?

This takes the swiss map implementation from github.com/dolthub/swiss plus some PRs I sent there to avoid having to pre-fill the internal datastructures with `empty` values, and copies that to our `tsdb` package making non-generic adjustments for our use case:
- empty value of the struct is valid, we allocate slices on first put.
- we already have a hash, use it
- store the hash where the original implementation stored the keys, and compare it after we found a hash suffix: so we avoid checking lset equality all the time
- only check lset equality when hash matches, accessing to memSeries internals (leveraging the fact that the implementation isn't generic)
- remove has(), clear(), etc. we don't need them.
- since we store the hash, rehashing is just allocating a new slice and copying the data.
- in this initial version, I didn't add the SIMD instructions, which would increase the performance on amd64.
  - [adding SIMD is trivial though](https://github.com/colega/prometheus/commit/5f600772f3bede13c623b0eda79f74ccf406eb42)

I didn't spend time on copying the tests from the original implementation: this structure is already covered by most of our tests, which should be enough to start reviewing. If we decide that we need them, I can bring them.

Benchmark results:
```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
                                             │    main     │               swapped               │
                                             │   sec/op    │   sec/op     vs base                │
HeadStripeSeriesCreate-12                      742.9n ± 5%   658.6n ± 1%  -11.35% (p=0.000 n=20)
HeadStripeSeriesGetExisting-12                 28.06n ± 0%   16.43n ± 0%  -41.46% (p=0.000 n=20)
HeadStripeSeriesCreateParallel-12              745.5n ± 2%   706.1n ± 3%   -5.29% (p=0.000 n=20)
HeadStripeSeriesCreate_PreCreationFailure-12   72.31n ± 0%   66.32n ± 0%   -8.28% (p=0.000 n=20)
geomean                                        183.1n        150.0n       -18.06%

                                             │     main      │               swapped               │
                                             │     B/op      │    B/op     vs base                 │
HeadStripeSeriesCreate-12                      658.5 ± 12%     639.5 ± 2%       ~ (p=1.000 n=20)
HeadStripeSeriesGetExisting-12                 0.000 ±  0%     0.000 ± 0%       ~ (p=1.000 n=20) ¹
HeadStripeSeriesCreateParallel-12              585.0 ±  1%     538.0 ± 0%  -8.03% (p=0.000 n=20)
HeadStripeSeriesCreate_PreCreationFailure-12   56.00 ±  0%     56.00 ± 0%       ~ (p=1.000 n=20) ¹
geomean                                                    ²               -2.79%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                             │     main     │               swapped               │
                                             │  allocs/op   │ allocs/op   vs base                 │
HeadStripeSeriesCreate-12                      3.000 ± 0%     3.000 ± 0%       ~ (p=1.000 n=20) ¹
HeadStripeSeriesGetExisting-12                 0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=20) ¹
HeadStripeSeriesCreateParallel-12              5.000 ± 0%     5.000 ± 0%       ~ (p=1.000 n=20) ¹
HeadStripeSeriesCreate_PreCreationFailure-12   3.000 ± 0%     3.000 ± 0%       ~ (p=1.000 n=20) ¹
geomean                                                   ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

Also, I took a profile of `BenchmarkHeadStripeSeriesCreate` with 50M series, and zooming into getOrSet:
![image](https://github.com/prometheus/prometheus/assets/1511481/ed1be45a-2547-4d20-8ac1-70131f67c65d)

Here it's seen that creating an entry in `map[ref]*memSeries` (`runtime.mapassign_fast64`) takes much longer than calling `seriesHashmap.set()`, while both things are called once from `getOrSet`.