@colega 2024-07-08
Unfortunately there aren't many metrics in prombench regarding the append time (or maybe they aren't on the dashboard?)