@colega 2024-07-09
Micro-benchmarks show an improvement, but my test (on a Mimir cluster) shows more time spent on `stripeSeries.getByRef` 🤷‍♂️