@colega 2024-07-11
I pushed an extra commit that uses higher bits for meta matching, instead of lower ones, as lower ones are always the same in a given stripe.

This improves the benchmarks, but I still don't see that improvement in production.